# swagger.yaml
swagger: "2.0"
info:
  title: "Book Finder"
  version: "1.0"
  contact: {}

definitions:
  model.Book:
    properties:
      author:
        type: string
      book_cover_url:
        type: string
      date_published:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object

paths:
  /:
    post:
      description: Initialize and configure HTTP handlers.
      responses:
        "204":
          description: No Content

  /books:
    get:
      description: Get a list of all books.
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/model.Book'
      summary: Get a list of all books

    post:
      consumes:
        - application/json
      description: Create a new book.
      parameters:
        - description: Book object to create
          in: body
          name: book
          required: true
          schema:
            $ref: '#/definitions/model.Book'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Book'
      summary: Create a new book

  /books/{id}:
    delete:
      description: Delete a book by its ID.
      parameters:
        - description: Book ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Book'
      summary: Delete a book by its ID

    get:
      description: Get a book by its ID.
      parameters:
        - description: Book ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Book'
      summary: Get a book by its ID

    put:
      consumes:
        - application/json
      description: Update a book by its ID.
      parameters:
        - description: Book ID
          in: path
          name: id
          required: true
          type: integer
        - description: Updated book object
          in: body
          name: book
          required: true
          schema:
            $ref: '#/definitions/model.Book'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Book'
      summary: Update a book by its ID